Test controller

+++ CalcolatriceScientifica/src/calcolatricescientifica/CommandDrop.java
@@ -19,7 +19,10 @@ public class CommandDrop implements Command{
     }
 
     @Override
-    public void esegui() {
+    public void esegui() throws Exception{
+        if(stack.lenght()==0){
+            throw new Exception();
+        }
         stack.drop();
     }
 }
