+++ CalcolatriceScientifica/src/calcolatricescientifica/CommandRadice.java
@@ -20,7 +20,10 @@ public class CommandRadice implements Command {
     }
 
     @Override
-    public void esegui() {
+    public void esegui() throws Exception{
+        if(stack.lenght()==0){
+            throw new Exception();
+        }
         NumeroComplesso nc = null;
         nc = stack.drop().radice();
         stack.push(nc);
